实作批次更新event功能

+++ app/controllers/admin/events_controller.rb
@@ -45,11 +45,19 @@ class Admin::EventsController < AdminController
     total = 0
     Array(params[:ids]).each do |event_id|
       event = Event.find(event_id)
-      event.destroy
-      total += 1
+      if params[:commit] == I18n.t(:bulk_update)
+        event.status = params[:event_status]
+        if event.save
+          total += 1
+        end
+        flash[:notice] = "成功更新 #{total} 笔资料"
+      elsif params[:commit] == I18n.t(:bulk_delete)
+        event.destroy
+        total += 1
+        flash[:alert] = "已删除 #{total} 笔资料"
+      end
     end
 
-    flash[:alert] = "成功完成 #{total} 笔"
     redirect_to admin_events_path
 
   end
